+++ tools/permutations.c
@@ -43,9 +43,9 @@ int main (int argc, char *argv[])
   cmd->string = strdup(argv[1]);
 
   struct graph *graph = graph_new();
-  struct cmd_token *token = new_cmd_token (START_TKN, cmd->attr, NULL, NULL);
+  struct cmd_token *token = cmd_token_new (START_TKN, cmd->attr, NULL, NULL);
   graph_new_node (graph, token, NULL);
-  command_parse_format (graph, cmd);
+  cmd_graph_parse (graph, cmd);
 
   permute (vector_slot (graph->nodes, 0));
 }
